Modified Classifier with seperated scraping (for Delhi University)
